+++ src/main/java/com/example/foodyrecommender/entity/Reservation.java
@@ -19,13 +19,13 @@ public class Reservation {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Long id;
+    private Integer id;
 
     @Column(name = "user_id")
-    private Long userId;
+    private Integer userId;
 
     @Column(name = "restaurant_id")
-    private Long restaurantId;
+    private Integer restaurantId;
 
     @Column(name = "booking_date")
     private LocalDate bookingDate;
